Vector-Embeddings/Images/streamlit-deployment/requirements.txt
numpy
pandas
streamlit
torch
Pillow
requests
scikit-learn
umap-learn
matplotlib
plotly
tqdm
transformers
datasets